Update simple push job
