+++ src/CloudFoundry.CloudController.V2.Client/CloudFoundry.CloudController.V2.Client.csproj
@@ -55,6 +55,7 @@
     <Compile Include="ClientExtensions\Data\DC_GetV1InfoResponse.cs" />
     <Compile Include="ClientExtensions\Data\DC_UpdateAppRequest.cs.cs" />
     <Compile Include="ClientExtensions\Info.cs" />
+    <Compile Include="ClientExtensions\ProgressEventArgs.cs" />
     <Compile Include="Client\Apps.cs" />
     <Compile Include="Client\AppUsageEvents.cs" />
     <Compile Include="Client\Buildpacks.cs" />
@@ -365,9 +366,8 @@
     <Compile Include="Metadata.cs" />
     <Compile Include="PageProperties.cs" />
     <Compile Include="PagedResponseCollection.cs" />
-    <Compile Include="ProgressInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PushJob.cs" />
+    <Compile Include="ClientExtensions\PushJob.cs" />
     <Compile Include="RequestOptions.cs" />
     <Compile Include="Utilities.cs" />
     <Compile Include="ZipUtil.cs" />
